+++ Packages/Style/Sources/Button/Types/ButtonStylePallete.swift
@@ -22,20 +22,88 @@ public struct ButtonStylePalette: Hashable, Sendable {
         self.backgroundPressed = backgroundPressed
         self.backgroundDisabled = backgroundDisabled
     }
+}
 
-    public static let primary = ButtonStylePalette(
+// MARK: - Static
+
+extension ButtonStylePalette {
+    static let primary = ButtonStylePalette(
         foreground: Colors.whiteSolid,
         foregroundPressed: Colors.whiteSolid,
         background: Colors.blue,
         backgroundPressed: Colors.blueDark,
         backgroundDisabled: Colors.blueDark.opacity(0.6)
     )
-    
-    public static let secondary = ButtonStylePalette(
+
+    static let secondary = ButtonStylePalette(
         foreground: Colors.whiteSolid,
         foregroundPressed: Colors.whiteSolid,
         background: Colors.blueSecondary,
         backgroundPressed: Colors.blueSecondaryHover,
         backgroundDisabled: Colors.blueSecondary.opacity(0.6)
+    )
+
+    static let blue = ButtonStylePalette(
+        foreground: Colors.whiteSolid,
+        foregroundPressed: Colors.whiteSolid,
+        background: Colors.blue,
+        backgroundPressed: Colors.blueDark,
+        backgroundDisabled: .clear
+    )
+
+    static let blueGrayPressed = ButtonStylePalette(
+        foreground: Colors.whiteSolid,
+        foregroundPressed: Colors.whiteSolid,
+        background: Colors.blue,
+        backgroundPressed: Colors.gray,
+        backgroundDisabled: .clear
+    )
+
+    static let gray = ButtonStylePalette(
+        foreground: Colors.whiteSolid,
+        foregroundPressed: Colors.whiteSolid,
+        background: Colors.grayLight,
+        backgroundPressed: Colors.gray,
+        backgroundDisabled: .clear
+    )
+
+    static let lightGray = ButtonStylePalette(
+        foreground: Colors.gray,
+        foregroundPressed: Colors.whiteSolid,
+        background: Colors.grayVeryLight,
+        backgroundPressed: Colors.grayLight.opacity(0.3),
+        backgroundDisabled: .clear
+    )
+
+    static let white = ButtonStylePalette(
+        foreground: Colors.whiteSolid,
+        foregroundPressed: Colors.whiteSolid,
+        background: Colors.white,
+        backgroundPressed: Colors.grayVeryLight,
+        backgroundDisabled: .clear
+    )
+
+    static let empty = ButtonStylePalette(
+        foreground: Colors.black,
+        foregroundPressed: Colors.black.opacity(0.5),
+        background: Colors.Empty.buttonsBackground,
+        backgroundPressed: Colors.Empty.buttonsBackground.opacity(0.5),
+        backgroundDisabled: .clear
+    )
+
+    static let amount = ButtonStylePalette(
+        foreground: Colors.black,
+        foregroundPressed: Colors.black.opacity(0.5),
+        background: Colors.grayVeryLight,
+        backgroundPressed: Colors.grayVeryLight.opacity(0.5),
+        backgroundDisabled: .clear
+    )
+
+    static let listStyleColor = ButtonStylePalette(
+        foreground: Colors.gray,
+        foregroundPressed: Colors.whiteSolid,
+        background: Colors.listStyleColor,
+        backgroundPressed: Colors.grayVeryLight,
+        backgroundDisabled: .clear
     )
 }
